% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceclogit.R
\name{niceclogit}
\alias{niceclogit}
\title{Lisa's Conditional Logistic Regression Table Function}
\usage{
niceclogit(df, fit, intercept = FALSE, ref = FALSE, labels = NA,
  blanks = FALSE, overallp = TRUE, est.dec = 2, ci.dec = 2,
  pval.dec = 3, htmlTable = FALSE, color = "#EEEEEE",
  printRMD = TRUE)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{clogit model object.}

\item{intercept}{If TRUE the intercept will be included in the table. Default is FALSE.}

\item{ref}{If TRUE, the reference category gets its own line (left blank). Default is FALSE.}

\item{labels}{Covariate labels - default is NA (variable names are used).}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{overallp}{If TRUE, a likelihood ratio test pvalue (using drop1 Chisq tests) will be calculated for each variable. Default is TRUE.}

\item{est.dec}{Number of decimal places for estimates - default is 2.}

\item{ci.dec}{Number of decimal places for CI - default is 2.}

\item{pval.dec}{Number of decimal places for pvalues - default is 3.}

\item{htmlTable}{Whether to use htmlTable package to display table (instead of xtable). Default = FALSE.}

\item{color}{Hex color to use for htmlTable output. Default = "#EEEEEE" (grey).}

\item{printRMD}{Whether to print resulting table to Rmd via xtable. Default = TRUE.}
}
\description{
This function creates a nice looking regression table from a clogit object. Estimates are presented as OR's (coefficients and CI's are exponentiated).
}
\keyword{conditional}
\keyword{glm}
\keyword{logistic}
\keyword{regression}
\keyword{table}
