% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicegee.R
\name{nicegee}
\alias{nicegee}
\title{Lisa's GEE Regression Table Function - fit must be geeglm type model}
\usage{
nicegee(df, fit, family = "gaussian", id, corstr = "exch",
  intercept = FALSE, ref = FALSE, labels = NA, blanks = FALSE,
  exp = FALSE, overallp = FALSE, est.dec = 2, ci.dec = 2,
  pval.dec = 3, estname = NA)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{geeglm (geepack) model object.}

\item{family}{Model family in quotes ("guassian", "binomial", "poisson").}

\item{id}{id variable.}

\item{corstr}{corstr type ("exch" is default).}

\item{ref}{If TRUE, the reference category gets its own line (left blank). 
Default is FALSE.}

\item{labels}{Covariate labels - default is NA.}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{exp}{Exponentiate estimates and CI's? (T or F).}

\item{overallp}{is whether to do overall Chisq test per variable}

\item{est.dec}{Number of decimal places for OR estimates - default is 4.}

\item{ci.dec}{Number of decimal places for 95% CI - default is 4.}

\item{pval.dec}{Number of decimal places for pvalues - default is 4.}

\item{estname}{Option to override estimate column name. Default is NA.}
}
\description{
Lisa's GEE Regression Table Function - fit must be geeglm type model
}
\keyword{gee}
\keyword{geeglm}
\keyword{geepack}
\keyword{logistic}
\keyword{pretty}
\keyword{regression}
\keyword{table}
