% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nicegeelrtable}
\alias{nicegeelrtable}
\title{Lisa's GEE Logistic Regression Table Function - fit must be geeglm type model}
\usage{
nicegeelrtable(df, fit, intercept = FALSE, ref = FALSE, labels = NA, id,
  blanks = FALSE, overallp = FALSE, est.dec = 4, ci.dec = 4,
  pval.dec = 3)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{Logistic Regression geeglm (geepack) model object.}

\item{ref}{If TRUE, the reference category gets its own line (left blank).
Default is FALSE.}

\item{labels}{Covariate labels - default is NA.}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{overallp}{is whether to do overall Chisq test per variable}

\item{est.dec}{Number of decimal places for OR estimates - default is 4.}

\item{ci.dec}{Number of decimal places for 95% CI - default is 4.}

\item{pval.dec}{Number of decimal places for pvalues - default is 4.}
}
\description{
Lisa's GEE Logistic Regression Table Function - fit must be geeglm type model
}
\keyword{gee}
\keyword{geeglm}
\keyword{geepack}
\keyword{logistic}
\keyword{pretty}
\keyword{regression}
\keyword{table}

