% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicegeepoistable.R
\name{nicegeepoistable}
\alias{nicegeepoistable}
\title{Lisa's GEE Poisson Regression Table Function - fit must be geeglm type model}
\usage{
nicegeepoistable(df, fit, intercept = FALSE, ref = FALSE, labels = NA,
  blanks = FALSE, overallp = FALSE, est.dec = 4, ci.dec = 4,
  pval.dec = 3)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{Poisson geeglm (geepack) model object.}

\item{ref}{If TRUE, the reference category gets its own line (left blank). Default is FALSE.}

\item{labels}{Covariate labels - default is NA.}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{overallp}{is FALSE always (LR p-value not supported in geeglm)}

\item{est.dec}{Number of decimal places for OR estimates - default is 4.}

\item{ci.dec}{Number of decimal places for 95% CI - default is 4.}

\item{pval.dec}{Number of decimal places for pvalues - default is 4.}
}
\description{
Lisa's GEE Poisson Regression Table Function - fit must be geeglm type model
}
\keyword{gee}
\keyword{geeglm}
\keyword{geepack}
\keyword{poisson}
\keyword{pretty}
\keyword{regression}
\keyword{table}

