% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceglmer.R
\name{niceglmer}
\alias{niceglmer}
\title{Lisa's GLMER Regression Table Function}
\usage{
niceglmer(df, fit, covs, family = "gaussian", exp = NA, estname = NA,
  intercept = FALSE, ref = FALSE, labels = NA, blanks = FALSE,
  overallp = FALSE, est.dec = 2, ci.dec = 2, pval.dec = 3,
  htmlTable = FALSE, color = "#EEEEEE", printRMD = TRUE)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{GLMER binomial model object.}

\item{covs}{Vector of covariate names in model.}

\item{family}{Model family in quotes ("guassian", "binomial", "poisson").}

\item{exp}{Option to exponentiate coefficients and CI's. Default is NA (exp for logistic and poisson families only).}

\item{estname}{Option to override estimate column name. Default is NA.}

\item{intercept}{If TRUE the intercept will be included in the table. Default is FALSE.}

\item{ref}{If TRUE, the reference category gets its own line (left blank). Default is FALSE.}

\item{labels}{Covariate labels - default is NA (variable names are used).}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{overallp}{If TRUE, a likelihood ratio test pvalue (using drop1 Chisq tests) will be calculated for each variable. Default is TRUE.}

\item{est.dec}{Number of decimal places for estimates - default is 2.}

\item{ci.dec}{Number of decimal places for CI - default is 2.}

\item{pval.dec}{Number of decimal places for pvalues - default is 3.}

\item{htmlTable}{Whether to use htmlTable package to display table (instead of xtable). Default = FALSE.}

\item{color}{Hex color to use for htmlTable output. Default = "#EEEEEE" (grey).}

\item{printRMD}{Whether to print resulting table to Rmd via xtable. Default = TRUE.}
}
\description{
This function creates a nice looking regression table from a glmer object
}
\keyword{effects}
\keyword{glmer}
\keyword{lmer}
\keyword{mixed}
\keyword{nicetable}
\keyword{regression}

