% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nicelinregtable}
\alias{nicelinregtable}
\title{Lisa's Linear Regression Table Function}
\usage{
nicelinregtable(df, fit, intercept = FALSE, ref = FALSE, labels = NA,
  blanks = FALSE, overallp = TRUE, est.dec = 2, ci.dec = 2,
  pval.dec = 3)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{lm model object.}

\item{intercept}{If TRUE the intercept will be included in the table. Default is FALSE.}

\item{ref}{If TRUE, the reference category gets its own line (left blank). Default is FALSE.}

\item{labels}{Covariate labels - default is NA (variable names are used).}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{overallp}{If TRUE, a likelihood ratio test pvalue (using drop1 Chisq tests) will be calculated for each variable. Default is TRUE.}

\item{est.dec}{Number of decimal places for estimates - default is 2.}

\item{ci.dec}{Number of decimal places for CI - default is 2.}

\item{pval.dec}{Number of decimal places for pvalues - default is 3.}
}
\description{
This function creates a nice looking regression table from a lm object
}
\keyword{linear}
\keyword{lm}
\keyword{regression}
\keyword{table}

