% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicepoistable.R
\name{nicepoistable}
\alias{nicepoistable}
\title{Lisa's Poisson Regression Table Function}
\usage{
nicepoistable(df, fit, intercept = FALSE, ref = FALSE, labels = NA,
  blanks = FALSE, overallp = TRUE, est.dec = 2, ci.dec = 2,
  pval.dec = 3)
}
\arguments{
\item{df}{Dataframe.}

\item{fit}{GLM model object (family = "poisson").}

\item{intercept}{If TRUE the intercept will be included in the table. Default is FALSE.}

\item{ref}{If TRUE, the reference category gets its own line (left blank). Default is FALSE.}

\item{labels}{Covariate labels - default is NA (variable names are used).}

\item{blanks}{If TRUE, blank lines will be inserted separating covariates - default is FALSE.}

\item{overallp}{If TRUE, a likelihood ratio test pvalue (using drop1 Chisq tests) will be calculated for each variable. Default is TRUE.}

\item{est.dec}{Number of decimal places for estimates - default is 2.}

\item{ci.dec}{Number of decimal places for CI - default is 2.}

\item{pval.dec}{Number of decimal places for pvalues - default is 3.}
}
\description{
This function re-formats the summary table from Poisson glm models 
to look nicer. Estimates and CI's are transformed to rate ratios.
Likelihood ratio p-values are added to the table for each covariate (optional).
The function returns a data frame and prints an html version 
for R markdown documents.
}
\keyword{glm}
\keyword{poisson}
\keyword{regression}
\keyword{table}

